# frontend/openapi.json
{
 "components": {
  "schemas": {
...
   "ChangePasswordRequest": {
    "additionalProperties": false,
    "properties": {
     "new_password": {
      "maxLength": 128,
      "minLength": 8,
      "title": "New Password",
      "type": "string"
     },
     "old_password": {
      "maxLength": 128,
      "minLength": 1,
      "title": "Old Password",
      "type": "string"
     }
    },
    "required": [
     "old_password",
     "new_password"
    ],
    "title": "ChangePasswordRequest",
    "type": "object"
   },
   "ChangePasswordResponse": {
    "properties": {
     "other_sessions_revoked": {
      "title": "Other Sessions Revoked",
      "type": "integer"
     }
    },
    "required": [
     "other_sessions_revoked"
    ],
    "title": "ChangePasswordResponse",
    "type": "object"
   },
...
   "DemoLoginRequest": {
    "additionalProperties": false,
    "properties": {
     "role": {
      "$ref": "#/components/schemas/UserRole"
     }
    },
    "required": [
     "role"
    ],
    "title": "DemoLoginRequest",
    "type": "object"
   },
...
   "EventIn": {
    "properties": {
     "district": {
      "anyOf": [
       {
        "maxLength": 80,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "District"
     },
     "ends_at": {
      "format": "date-time",
      "title": "Ends At",
      "type": "string"
     },
     "intensity": {
      "default": 1.0,
      "exclusiveMinimum": 0.0,
      "maximum": 10.0,
      "title": "Intensity",
      "type": "number"
     },
     "name_bn": {
      "maxLength": 120,
      "minLength": 1,
      "title": "Name Bn",
      "type": "string"
     },
     "name_en": {
      "maxLength": 120,
      "minLength": 1,
      "title": "Name En",
      "type": "string"
     },
     "starts_at": {
      "format": "date-time",
      "title": "Starts At",
      "type": "string"
     },
     "type": {
      "$ref": "#/components/schemas/EventType"
     }
    },
    "required": [
     "type",
     "name_en",
     "name_bn",
     "starts_at",
     "ends_at"
    ],
    "title": "EventIn",
    "type": "object"
   },
   "EventItem": {
    "properties": {
     "district": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "District"
     },
     "ends_at": {
      "format": "date-time",
      "title": "Ends At",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "intensity": {
      "title": "Intensity",
      "type": "number"
     },
     "name_bn": {
      "title": "Name Bn",
      "type": "string"
     },
     "name_en": {
      "title": "Name En",
      "type": "string"
     },
     "starts_at": {
      "format": "date-time",
      "title": "Starts At",
      "type": "string"
     },
     "type": {
      "$ref": "#/components/schemas/EventType"
     }
    },
    "required": [
     "id",
     "type",
     "name_en",
     "name_bn",
     "starts_at",
     "ends_at",
     "district",
     "intensity"
    ],
    "title": "EventItem",
    "type": "object"
   },
   "EventPage": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/EventItem"
      },
      "title": "Items",
      "type": "array"
     },
     "page": {
      "title": "Page",
      "type": "integer"
     },
     "page_size": {
      "title": "Page Size",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "EventPage",
    "type": "object"
   },
   "EventType": {
    "enum": [
     "salary",
     "eid",
     "hat_bazar",
     "weather",
     "holiday"
    ],
    "title": "EventType",
    "type": "string"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Lang": {
    "enum": [
     "bn",
     "en"
    ],
    "title": "Lang",
    "type": "string"
   },
...
   "LoginRequest": {
    "additionalProperties": false,
    "properties": {
     "email": {
      "maxLength": 254,
      "minLength": 3,
      "title": "Email",
      "type": "string"
     },
     "password": {
      "maxLength": 128,
      "minLength": 1,
      "title": "Password",
      "type": "string"
     }
    },
    "required": [
     "email",
     "password"
    ],
    "title": "LoginRequest",
    "type": "object"
   },
...
   "Theme": {
    "enum": [
     "light",
     "dark",
     "system"
    ],
    "title": "Theme",
    "type": "string"
   },
   "TokenResponse": {
    "description": "The refresh token is never in the body: it travels only in the httpOnly cookie.",
    "properties": {
     "access_token": {
      "title": "Access Token",
      "type": "string"
     },
     "expires_in": {
      "title": "Expires In",
      "type": "integer"
     },
     "token_type": {
      "const": "bearer",
      "default": "bearer",
      "title": "Token Type",
      "type": "string"
     },
     "user": {
      "$ref": "#/components/schemas/UserOut"
     }
    },
    "required": [
     "access_token",
     "expires_in",
     "user"
    ],
    "title": "TokenResponse",
    "type": "object"
   },
...
   "UserOut": {
    "properties": {
     "agent_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id"
     },
     "avatar_color": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avatar Color"
     },
     "digits": {
      "$ref": "#/components/schemas/Lang"
     },
     "display_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     },
     "distributor_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Distributor Id"
     },
     "email": {
      "title": "Email",
      "type": "string"
     },
     "full_name": {
      "title": "Full Name",
      "type": "string"
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "lang": {
      "$ref": "#/components/schemas/Lang"
     },
     "last_login_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Login At"
     },
     "notify_in_app": {
      "title": "Notify In App",
      "type": "boolean"
     },
     "role": {
      "$ref": "#/components/schemas/UserRole"
     },
     "theme": {
      "$ref": "#/components/schemas/Theme"
     },
     "tour_done": {
      "title": "Tour Done",
      "type": "boolean"
     }
    },
    "required": [
     "id",
     "email",
     "full_name",
     "role",
     "agent_id",
     "distributor_id",
     "lang",
     "theme",
     "digits",
     "notify_in_app",
     "tour_done",
     "display_name",
     "avatar_color",
     "last_login_at"
    ],
    "title": "UserOut",
    "type": "object"
   },
   "UserRole": {
    "enum": [
     "agent",
     "distributor",
     "admin"
    ],
    "title": "UserRole",
    "type": "string"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "title": "AgentPulse AI",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/auth/change-password": {
   "post": {
    "operationId": "change_password_api_v1_auth_change_password_post",
    "parameters": [
     {
      "in": "cookie",
      "name": "ap_refresh",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 512,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Ap Refresh"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChangePasswordRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChangePasswordResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Change Password",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/auth/demo-login": {
   "post": {
    "description": "One-click sign-in as a seeded is_demo account (DEMO_MODE only, rate-limited per IP,\naudited). Never enable DEMO_MODE on a public deployment.",
    "operationId": "demo_login_api_v1_auth_demo_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DemoLoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Demo Login",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/auth/login": {
   "post": {
    "operationId": "login_api_v1_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Login",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/auth/logout": {
   "post": {
    "description": "Works with an expired access token: the cookie alone identifies the session.",
    "operationId": "logout_api_v1_auth_logout_post",
    "parameters": [
     {
      "in": "cookie",
      "name": "ap_refresh",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 512,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Ap Refresh"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Logout",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/auth/me": {
   "get": {
    "operationId": "me_api_v1_auth_me_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Me",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/auth/refresh": {
   "post": {
    "operationId": "refresh_api_v1_auth_refresh_post",
    "parameters": [
     {
      "in": "cookie",
      "name": "ap_refresh",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 512,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Ap Refresh"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Refresh",
    "tags": [
     "auth"
    ]
   }
  },
...
  "/api/v1/events": {
   "get": {
    "description": "Events overlapping [from, to), by start time; `district` keeps nationwide events too.",
    "operationId": "list_events_api_v1_events_get",
    "parameters": [
     {
      "in": "query",
      "name": "from",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date-time",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "From"
      }
     },
     {
      "in": "query",
      "name": "to",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date-time",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "To"
      }
     },
     {
      "in": "query",
      "name": "type",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/EventType"
        },
        {
         "type": "null"
        }
       ],
       "title": "Type"
      }
     },
     {
      "in": "query",
      "name": "district",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 80,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "District"
      }
     },
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "maximum": 10000,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 50,
       "maximum": 200,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventPage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Events",
    "tags": [
     "events"
    ]
   },
   "post": {
    "description": "Add an event (audit_log). Forecasts use it from the next precompute.",
    "operationId": "create_event_api_v1_events_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EventIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventItem"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Create Event",
    "tags": [
     "events"
    ]
   }
  },
  "/api/v1/events/{event_id}": {
   "delete": {
    "description": "Delete an event (audit_log keeps the deleted row).",
    "operationId": "delete_event_api_v1_events__event_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "event_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Event Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Delete Event",
    "tags": [
     "events"
    ]
   },
   "put": {
    "description": "Replace an event (audit_log keeps before/after).",
    "operationId": "update_event_api_v1_events__event_id__put",
    "parameters": [
     {
      "in": "path",
      "name": "event_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Event Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EventIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventItem"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Update Event",
    "tags": [
     "events"
    ]
   }
  },
...
 }
}